requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
streamlit==1.44.1
openai==1.69.0
python-dotenv==1.1.0
pymongo==4.11.3
pydantic==2.11.1
requests==2.31.0
python-multipart==0.0.6
langgraph==0.4.5
langgraph-checkpoint-mongodb==0.1.3